Fixed Multiple Block Problem
I had to change the way push potions are scripted so that it is more readable, and I used joints to connect the wooden blocks so that now the larger block does not need a collider.  Also, The push potion should just apply a force on everything, becuase anything that shouldn't move wont have a rigidbody.